// utilities/types.tsx
export interface User {
    EmployeeId: number;
    Username: string;
    UserId: string;
    Email: string;
    Role: string;
    Department: string | null;
    PhoneNumber: string | null;
    DateOfJoining: string | null;
    Address: string | null;
    ImagePath: string | null;
    CreatedOn: string;
    ModifiedOn: string | null;
  }
  